make receipts test stand alone Fixes #707
